Rename cloud_omezarr --> omezarr
