+++ src/main/resources/META-INF/services/qupath.lib.gui.extensions.QuPathExtension
@@ -1,1 +1,1 @@
-dimilab.qupath.ext.cloud_omezarr.CloudOmeZarrExtension
+dimilab.qupath.ext.omezarr.CloudOmeZarrExtension
