+++ src/main/resources/META-INF/services/qupath.lib.images.servers.ImageServerBuilder
@@ -1,1 +1,1 @@
-dimilab.qupath.ext.cloud_omezarr.CloudOmeZarrServerBuilder
+dimilab.qupath.ext.omezarr.CloudOmeZarrServerBuilder
