+++ src/main/resources/dimilab/qupath/ext/omezarr/ui/interface.fxml
