+++ src/main/resources/dimilab/qupath/ext/omezarr/ui/strings.properties
