+++ src/test/kotlin/dimilab/qupath/ext/omezarr/OmeXmlUtilsTest.kt
@@ -1,4 +1,4 @@
-package dimilab.qupath.ext.cloud_omezarr
+package dimilab.qupath.ext.omezarr
 
 import kotlin.io.path.toPath
 import kotlin.test.Test
